Fixed issue where the modify script for raw images could not be executed
As part of this, I refactored the writeBuildDirFile and writeCombustionDirFile methods to return the full name of the file written. They are meant to be utility functions to prevent the caller from having to manually assemble the full paths to common locations, and it makes sense that these calls would also tell the user what the filename ended up being.