Basic infrastructure for user-facing audit messages (#49)
Adds in auditing ability (separate from logging) specifically for handling user-readable messages we want to display to the user. This PR also includes usages of these in each of the relevant components as well as some basic audit information about the rest of the build process. However, this wasn't a full review of what/when we should output to the user, which will need further work in the future.