Add a Network validation section for APIVIP sanity checks
Almost all sections within the image definition go through sanity checks
before actually creating a new image. However, EIB misses a dedicated
section to test the Network part of the Kubernetes configuration.

Add a validateNetwork function to in order to make sure that the specified
APIVIP value is well-formed and valid, regardless of wheather it is in
IPv4 or IPv6 format.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>