Fix the VIP URL generation for IPv6
When configuring Kubernetes with multiple nodes, EIB requires an IP address
to be used as a target for load balancing purposes. However, it assumes this
address to be in an IPv4 format, resulting in a broken mangling when IPv6 is
used instead.

Correct this problem by letting netip package do the handling of IP addresses,
hiding version specific issues. This also allows for easier and additional
input validation with minimal effort.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>