Create the required Elemental Agent directory structure during Combustion (#634)
In this PR we create the required /etc/rancher/elemental/agent directory
structure during the initial Elemental setup code. A config file is checked
within the systemd startup condition, and it seemingly fails if this directory
isn't already in-place. This seems to only be an issue in SL Micro 6.1, but
it shouldn't harm for previous versions of Micro.

Fixes #632